it works

+++ extensions/guild/moderation.py
@@ -1,6 +1,7 @@
 import asyncio
 
 import discord
+from discord import DMChannel
 from discord.ext import commands
 
 from config import config
@@ -13,6 +14,11 @@ class EventsModeration(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message):
+        if isinstance(message.channel, DMChannel):
+            print("Сообщение в DM!")
+            print(message)
+            print(message.channel)
+            return
         if message.channel.category_id == forum_tasks_id and message.content.startswith("(("):
             await asyncio.sleep(20)
             await message.delete()
